Feat/토글 달기 불필요한  로직 간단히 수정
