+++ src/apis/meeting/meetingTypes.ts
@@ -45,3 +45,10 @@ export interface MeetingCreateResponse {
   meetingId: string;
   createdAt: string;
 }
+
+export interface Anchor {
+  top: number;
+  left: number;
+  right: number;
+  bottom: number;
+}
