+++ src/pages/meeting/detail/MeetingDetailPage.module.scss
@@ -51,26 +51,3 @@
   color: white;
 }
 
-.dropdown{
-  position:fixed;
-  background-color: aqua;
-  width: 15%;
-  height: 5%;
-  top: 3rem;
-  right: 0.8rem;
-  display: flex;
-  align-items: center;
-  border-radius: 16px;
-  background: #FFFFFFB2;
-  backdrop-filter: blur(50px);
-
-  box-shadow: 0px 0px 20px 0px #00000014;
-
-  &__deleteBtn{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 0.8rem;
-    gap: 0.9rem;
-  }
-}
